+++ packages/signers/signer/src/from-raw-signer.ts
@@ -30,7 +30,7 @@ export function getPolkadotSigner(
     const decMeta = unifyMetadata(decAnyMetadata(metadata))
     const extra: Array<Uint8Array> = []
     const additionalSigned: Array<Uint8Array> = []
-    decMeta.extrinsic.signedExtensions.map(({ identifier }) => {
+    decMeta.extrinsic.signedExtensions[0].map(({ identifier }) => {
       const signedExtension = signedExtensions[identifier]
       if (!signedExtension)
         throw new Error(`Missing ${identifier} signed extension`)