+++ packages/compare-runtimes/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+- Extensions are now compared by `ExtVersion` instead of individually
+
 ## 0.2.3 to 0.2.13 - 2026-01-05
 
 ### Fixed