+++ packages/signers/pjs-signer/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Fixed
+
+- Use right signed extensions for transaction building
+
 ## 0.6.1 to 0.6.18 - 2026-01-05
 
 ### Fixed